+++ src/pages/_error.tsx
@@ -4,7 +4,7 @@ import NextError, { ErrorProps as NextErrorProps } from "next/error";
 import React from "react";
 
 import { ErrorFallbackComponent } from "@src/components/error-fallback-component/error-fallback-component";
-import { Layout } from "@src/components/layout-remove";
+import { Layout } from "@src/components/layout";
 
 type CustomErrorComponentProps = {
   error: Error;
