+++ src/pages/account/logins/[type]/[id].tsx
@@ -7,7 +7,7 @@ import { UserCookie } from "@src/@types/user-cookie";
 import { NoDataReturned, NoIdentityFound } from "@src/client-errors";
 import { Container } from "@src/components/containers/container";
 import { IdentityOverview } from "@src/components/identity-overview/identity-overview";
-import { Layout } from "@src/components/layout-remove";
+import { Layout } from "@src/components/layout";
 import { NavigationBreadcrumbs } from "@src/components/navigation-breadcrumbs/navigation-breadcrumbs";
 import { GraphqlErrors } from "@src/errors";
 import { withAuth } from "@src/middlewares/with-auth";
