+++ src/pages/account/logins/[type]/new.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import { IdentityTypes } from "@lib/@types";
 import { AddIdentityForm } from "@src/components/add-identity-form/add-identity-form";
 import { Container } from "@src/components/containers/container";
-import { Layout } from "@src/components/layout-remove";
+import { Layout } from "@src/components/layout";
 import { NavigationBreadcrumbs } from "@src/components/navigation-breadcrumbs/navigation-breadcrumbs";
 import { withAuth } from "@src/middlewares/with-auth";
 import { withLogger } from "@src/middlewares/with-logger";
