+++ src/pages/account/logins/[type]/validation/[eventId].tsx
@@ -4,7 +4,7 @@ import React from "react";
 
 import { IdentityTypes } from "@lib/@types/Identity";
 import { Container } from "@src/components/containers/container";
-import { Layout } from "@src/components/layout-remove";
+import { Layout } from "@src/components/layout";
 import { NavigationBreadcrumbs } from "@src/components/navigation-breadcrumbs/navigation-breadcrumbs";
 import { ValidateIdentityForm } from "@src/components/validate-identity-form/validate-identity-form";
 import { withAuth } from "@src/middlewares/with-auth";
